chatbot-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
python-dotenv==1.0.0

# Google Gemini
google-generativeai==0.8.3

# Utils
httpx==0.27.0
